+++ libraries/aircraft-simulator/src/aircraft.cpp
@@ -63,7 +63,7 @@ void Aircraft::checkAcceleration()
     // }
 }
 
-long double Aircraft::sendSignal(Tower tower, const OurVector<3>& current_position)
+double Aircraft::sendSignal(Tower tower, const OurVector<3>& current_position)
 {
     return tower.calculateTime(current_position);
 }
