+++ libraries/aircraft-simulator/src/plotter.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <plotter.h>
 
-void Plotter::addPoint(long double x, long double y, long double z)
+void Plotter::addPoint(double x, double y, double z)
 {
 	assert(_file->is_open());
 	(*_file) << x << ' ' << y << ' ' << z << std::endl;
